+++ util/fzf.go
@@ -5,9 +5,9 @@ import (
 	"github.com/junegunn/fzf/src/util"
 )
 
-func FuzzyScore(input, target string) int {
+func FuzzyScore(input, target string) float64 {
 	chars := util.ToChars([]byte(target))
 	res, _ := algo.FuzzyMatchV2(false, true, true, &chars, []rune(input), true, nil)
 
-	return res.Score
+	return float64(res.Score)
 }
